+++ pykeg/__init__.py
@@ -1,5 +1,5 @@
 # Package version.
-__version__ = '0.9.12'
+__version__ = '0.9.13-pre'
 
 # Internal version number, bumped every time "kb_upgrade" is needed.
 EPOCH = 103